requirements.txt
streamlit
web3
transformers
torch
python-dotenv